pyproject.toml
[project]
name = "bwim-purple-agent"
version = "0.1.0"
description = "LLM-powered builder agent for Build What I Mean benchmark"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.5",
    "anthropic[bedrock]>=0.49.0",
    "pydantic>=2.11.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.35.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
